Added eventdetailscopy
Event details copy is a page to fuck around with the grabdetails method.

Basically the grabdetails (line 197) function returns undefined. when I call it for console logging on line 184. I want it to return the array of JSON objects but I cannot for the life of me figure out how to do this.